+++ src/lpc/mod.rs
@@ -8,6 +8,9 @@ use structopt::StructOpt;
 use ecoz2_lib::lpc_signals;
 use utl;
 
+mod libpar;
+mod lpc_rs;
+
 #[derive(StructOpt, Debug)]
 pub struct LpcOpts {
     /// Prediction order
@@ -35,6 +38,14 @@ pub struct LpcOpts {
     /// all `.wav` under them will be used.
     #[structopt(parse(from_os_str))]
     sgn_filenames: Vec<PathBuf>,
+
+    /// Use Rust implementation
+    #[structopt(long)]
+    zrs: bool,
+
+    /// Use Rust "parallel" implementation
+    #[structopt(long)]
+    zrsp: bool,
 }
 
 pub fn main(opts: LpcOpts) {
@@ -53,18 +64,71 @@ pub fn main_lpc(opts: LpcOpts) -> Result<(), Box<dyn Error>> {
         minpc,
         split,
         sgn_filenames,
+        zrs,
+        zrsp,
     } = opts;
 
     let actual_sgn_filenames = utl::get_actual_filenames(sgn_filenames, ".wav", "signals")?;
 
-    lpc_signals(
-        prediction_order,
-        window_length_ms,
-        offset_length_ms,
-        minpc,
-        split,
-        actual_sgn_filenames,
-    );
+    if zrs {
+        assert!(!zrsp);
+        main_lpc_rs(
+            actual_sgn_filenames,
+            prediction_order,
+            window_length_ms,
+            offset_length_ms,
+        );
+    } else if zrsp {
+        main_lpc_par_rs(
+            actual_sgn_filenames,
+            prediction_order,
+            window_length_ms,
+            offset_length_ms,
+        );
+    } else {
+        lpc_signals(
+            prediction_order,
+            window_length_ms,
+            offset_length_ms,
+            minpc,
+            split,
+            actual_sgn_filenames,
+        );
+    }
 
     Ok(())
+}
+
+fn main_lpc_rs(
+    sgn_filenames: Vec<PathBuf>,
+    prediction_order: usize,
+    window_length_ms: usize,
+    offset_length_ms: usize,
+) {
+    for sgn_filename in sgn_filenames {
+        lpc_rs::lpc_rs(
+            sgn_filename,
+            None,
+            prediction_order,
+            window_length_ms,
+            offset_length_ms,
+        );
+    }
+}
+
+fn main_lpc_par_rs(
+    sgn_filenames: Vec<PathBuf>,
+    prediction_order: usize,
+    window_length_ms: usize,
+    offset_length_ms: usize,
+) {
+    for sgn_filename in sgn_filenames {
+        libpar::lpc_par(
+            sgn_filename,
+            None,
+            prediction_order,
+            window_length_ms,
+            offset_length_ms,
+        );
+    }
 }