+++ src/prd/mod.rs
@@ -1,6 +1,9 @@
+extern crate serde;
 extern crate structopt;
 
 use std::error::Error;
+use std::fs::File;
+use std::io::{BufReader, BufWriter};
 use std::path::PathBuf;
 
 use structopt::StructOpt;
@@ -8,6 +11,7 @@ use structopt::StructOpt;
 use ecoz2_lib::prd_show_file;
 
 use self::EcozPrdCommand::Show;
+use itertools::enumerate;
 
 #[derive(StructOpt, Debug)]
 pub struct PrdMainOpts {
@@ -39,6 +43,10 @@ pub struct PrdShowOpts {
     /// File to read
     #[structopt(parse(from_os_str))]
     file: PathBuf,
+
+    /// Use Rust implementation
+    #[structopt(long)]
+    zrs: bool,
 }
 
 pub fn main(opts: PrdMainOpts) {
@@ -57,9 +65,63 @@ pub fn prd_show(opts: PrdShowOpts) -> Result<(), Box<dyn Error>> {
         from,
         to,
         file,
+        zrs,
     } = opts;
 
-    prd_show_file(file, show_reflection, from, to);
+    if zrs {
+        println!("warn: ignoring any options in rust impl.");
+        prd_show_rs(file);
+    } else {
+        prd_show_file(file, show_reflection, from, to);
+    }
 
     Ok(())
+}
+
+// NOTE: for Rust implementation (preliminary)
+
+fn prd_show_rs(prd_filename: PathBuf) {
+    let mut prd = load(prd_filename.to_str().unwrap()).unwrap();
+    prd.show();
+}
+
+#[derive(serde::Serialize, serde::Deserialize, Debug)]
+pub struct Predictor {
+    pub class_name: String,
+    pub prediction_order: usize,
+    pub vectors: Vec<Vec<f64>>,
+}
+
+impl Predictor {
+    pub fn save(&mut self, filename: &str) -> Result<(), Box<dyn Error>> {
+        let f = File::create(filename)?;
+        let bw = BufWriter::new(f);
+        serde_cbor::to_writer(bw, &self)?;
+        Ok(())
+    }
+
+    pub fn show(&mut self) {
+        println!(
+            " class_name = '{}' prediction_order: {} vectors: {}",
+            self.class_name,
+            self.prediction_order,
+            self.vectors.len(),
+        );
+        for (i, v) in enumerate(&self.vectors) {
+            print!("{:4}: ", i);
+            let mut comma = "";
+            for val in v {
+                print!("{}{}", comma, val);
+                comma = ", ";
+            }
+            println!();
+        }
+    }
+}
+
+pub fn load(filename: &str) -> Result<Predictor, Box<dyn Error>> {
+    let f = File::open(filename)?;
+    let br = BufReader::new(f);
+    let predictor = serde_cbor::from_reader(br)?;
+    Ok(predictor)
 }